pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
openai = "^1.3"
tiktoken = "^0.7.0"
retrying = "^1.3.4"
flask = "^3.0.3"
numpy = "^1.26.4"
sentence-transformers = "^3.0.0"
faiss-cpu = "^1.8.0"
torch = "^2.3.0"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"